% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignNewlineIndent.R
\name{assignNewlineIndent}
\alias{assignNewlineIndent}
\title{assignNewlineIndent}
\usage{
assignNewlineIndent()
}
\description{
Call this function as an addin to insert an assignment (<-) and newline at the cursor position with an extra indent level.
}

