% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeNewlineIndent.R
\name{pipeNewlineIndent}
\alias{pipeNewlineIndent}
\title{pipeNewlineIndent}
\usage{
pipeNewlineIndent()
}
\description{
Call this function as an addin to insert pipe and newline at the cursor position. Respects your indents.
}

