% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeNewlineIndent.R
\name{pipeNewlineIndentDplyr}
\alias{pipeNewlineIndentDplyr}
\title{pipeNewlineIndentDplyr}
\usage{
pipeNewlineIndentDplyr()
}
\description{
Call this function as an addin to insert pipe and newline and dplyr:: at the cursor position. Respects your indents.
}

