% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignNewlineIndent.R
\name{assignNewlineIndent}
\alias{assignNewlineIndent}
\title{Insert ' <- newline'.}
\usage{
assignNewlineIndent()
}
\description{
Call this function as an addin to insert and assignment and newline at the cursor position.
}

