% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeNewlineIndent.R
\name{pipeNewlineIndent}
\alias{pipeNewlineIndent}
\title{Insert ' %<% newline'.}
\usage{
pipeNewlineIndent()
}
\description{
Call this function as an addin to insert a pipe and newline at the cursor position.
}

